# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/iframe/alarms": {
      "get": {
        "description": "Returns an iFrame with alarms from multiple sources.",
        "produces": ["text/html"],
        "summary": "Alarms iFrame",
        "parameters": [
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "example": 5,
            "description": "Limits the number of items in the iFrame.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "example": "netdata,radarr,sonarr",
            "description": "Alarms to show. Available values: netdata, radarr, sonarr, prowlarr, speedtest-tracker, pihole, kavita, kaizoku",
            "name": "alarms",
            "in": "query",
            "required": true
          },
          {
            "type": "boolean",
            "example": false,
            "description": "Sort alarms in descending order. Defaults to false.",
            "name": "sort_desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/iframe/cinemark": {
      "get": {
        "description": "Returns an iFrame with the on display movies in specific Cinemark theaters. I recommend you to get the movies from the theaters of your city.",
        "produces": ["text/html"],
        "summary": "Cinemark Brazil iFrame",
        "parameters": [
          {
            "type": "string",
            "example": "715, 1222, 4555",
            "description": "The theater IDs to get movies from. It used to be easy to get, but now it's harder. To get it, you need to access the cinemark site, select a theater, open your browser developer console, go to the ",
            "name": "theaterIds",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "integer",
            "example": 5,
            "description": "Limits the number of items in the iFrame.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload.",
            "name": "api_url",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/iframe/linkwarden": {
      "get": {
        "description": "Returns an iFrame with Linkwarden bookmarks.",
        "produces": ["text/html"],
        "summary": "Linkwarden  bookmarks iFrame",
        "parameters": [
          {
            "type": "integer",
            "example": 1,
            "description": "Get bookmarks only from this collection. You can get the collection ID by going to the collection page. The ID should be on the URL. The ID of the default collection **Unorganized** is 1 because the URL is https://domain.com/collections/1.",
            "name": "collectionId",
            "in": "query"
          },
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "integer",
            "example": 5,
            "description": "Limits the number of items in the iFrame.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "top",
            "description": "Background position of each bookmark card. Use '%25' in place of '%', like '50%25 47.2%25' to get '50% 47.2%'. Defaults to 50% 47.2%.",
            "name": "background_position",
            "in": "query"
          },
          {
            "type": "string",
            "example": "cover",
            "description": "Background size of each bookmark card. Use '%25' in place of '%'. Defaults to cover.",
            "name": "background_size",
            "in": "query"
          },
          {
            "type": "string",
            "example": "blur(5px",
            "description": "Background filter of each bookmark card. Use '%25' in place of '%'. Defaults to brightness(0.3).",
            "name": "background_filter",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/iframe/media_releases": {
      "get": {
        "description": "Returns an iFrame with the media releases of today. The media releases are from Radarr/Sonarr.",
        "produces": ["text/html"],
        "summary": "Media Releases",
        "parameters": [
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload. Also used by the button to set the task done, if not provided, the button will not appear.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "physical",
            "description": "Filter movies get from Radarr. Can be 'inCinemas', 'physical', or 'digital'. Defaults to 'inCinemas'",
            "name": "radarrReleaseType",
            "in": "query"
          },
          {
            "type": "boolean",
            "example": true,
            "description": "Specify if show unmonitored media. Defaults to false.",
            "name": "showUnmonitored",
            "in": "query"
          },
          {
            "type": "boolean",
            "example": false,
            "description": "Specify if show the episodes' (Sonarr) release hour and minute. Defaults to true.",
            "name": "showEpisodesHour",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/iframe/media_requests": {
      "get": {
        "description": "Returns an iFrame with Overseerr and Jellyseerr media requests list. Returns all requests if the user's API token has the ADMIN or MANAGE_REQUESTS permissions. Otherwise, only the logged-in user's requests are returned.",
        "produces": ["text/html"],
        "summary": "Overseerr and Jellyseerr Media Requests",
        "parameters": [
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload. Also used by the button to set the task done, if not provided, the button will not appear.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "example": 5,
            "description": "Limits the number of items in the iFrame.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "example": "all",
            "description": "Available values : all, approved, available, pending, processing, unavailable, failed",
            "name": "filter",
            "in": "query"
          },
          {
            "type": "string",
            "example": "added",
            "description": "Available values : added, modified. Defaults to added",
            "name": "sort",
            "in": "query"
          },
          {
            "type": "string",
            "example": "1",
            "description": "If specified, only requests from that particular overseerr user ID will be returned.",
            "name": "requestedByOverseerr",
            "in": "query"
          },
          {
            "type": "string",
            "example": "1",
            "description": "If specified, only requests from that particular jellyseerr user ID will be returned.",
            "name": "requestedByJellyseerr",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/iframe/overseerr": {
      "get": {
        "description": "Returns an iFrame with Overseerr media requests list. Returns all requests if the user's API token has the ADMIN or MANAGE_REQUESTS permissions. Otherwise, only the logged-in user's requests are returned.",
        "produces": ["text/html"],
        "summary": "Overseerr Media Requests",
        "parameters": [
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload. Also used by the button to set the task done, if not provided, the button will not appear.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "example": 5,
            "description": "Limits the number of items in the iFrame.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "example": "all",
            "description": "Available values : all, approved, available, pending, processing, unavailable, failed",
            "name": "filter",
            "in": "query"
          },
          {
            "type": "string",
            "example": "added",
            "description": "Available values : added, modified. Defaults to added",
            "name": "sort",
            "in": "query"
          },
          {
            "type": "string",
            "example": "1",
            "description": "If specified, only requests from that particular user ID will be returned.",
            "name": "requestedBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/iframe/uptimekuma": {
      "get": {
        "description": "Returns an iFrame with Uptime Kuma sites overview.",
        "produces": ["text/html"],
        "summary": "Uptime Kuma iFrame",
        "parameters": [
          {
            "type": "string",
            "example": "uptime-kuma-slug",
            "description": "You need to create a status page in Uptime Kuma and select which sites/services this status page will show. While creating the status page, it'll request **you** to create a slug, after creating the status page, provide this slug here. This iFrame will show data only of the sites/services of this specific status page!",
            "name": "slug",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "light",
            "description": "Homarr theme, defaults to light. If it's different from your Homarr theme, the background turns white",
            "name": "theme",
            "in": "query"
          },
          {
            "type": "string",
            "example": "https://sub.domain.com",
            "description": "API URL used by your browser. Use by the iFrames to check any update, if there is an update, the iFrame reloads. If not specified, the iFrames will never try to reload.",
            "name": "api_url",
            "in": "query",
            "required": true
          },
          {
            "type": "boolean",
            "example": true,
            "description": "Show the title 'Uptime Kuma' on the iFrame.",
            "name": "showTitle",
            "in": "query"
          },
          {
            "type": "string",
            "example": "vertical",
            "description": "Orientation of the containers, defaults to horizontal.",
            "name": "orientation",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/iframe/vikunja/set_task_done": {
      "patch": {
        "description": "Set a Vikunja task as done.",
        "produces": ["application/json"],
        "summary": "Set Vikunja task done",
        "parameters": [
          {
            "type": "integer",
            "example": 1,
            "description": "The task ID.",
            "name": "taskId",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "example": 86400,
            "description": "The task Repeat After value.",
            "name": "taskRepeatAfter",
            "in": "query",
            "required": false
          },
          {
            "type": "integer",
            "example": 0,
            "description": "The task Repeat Mode value.",
            "name": "taskRepeatMode",
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Task done",
            "schema": {
              "$ref": "#/definitions/routes.messsageResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "routes.messsageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}